refactor typing tests
