+++ src/main/kotlin/org/move/utils/tests/MoveTypingTestCase.kt
@@ -1,10 +1,20 @@
 package org.move.utils.tests
 
 import org.intellij.lang.annotations.Language
+import org.move.lang.MoveFileType
 
 abstract class MvTypingTestCase : MvTestBase() {
     protected fun doTest(c: Char = '\n') = checkByFile {
         myFixture.type(c)
+    }
+
+    protected fun doTest(@Language("Move") before: String, type: Char, @Language("Move") after: String) {
+        val beforeText = replaceCaretMarker(before)
+        val afterText = replaceCaretMarker(after)
+
+        myFixture.configureByText(MoveFileType, beforeText)
+        myFixture.type(type)
+        myFixture.checkResult(afterText)
     }
 
     protected fun doTestByText(
